3/27 Update
Added "Hit" and "Stay" buttons -- Having trouble with Event Listeners on these... class methods require arguments and this breaks the "onclick" functionality as JS fires the method on page load...

Reorganized some methods to make more sense... still having trouble deciding where to add methods "where it makes sense"

Added the score display. It tallies up the cards that are initially dealt to the player. This is just a test function but the inner code is pretty simple and should be easy to implement on event listeners.