3/29 Update
Got some of the functionality of the winner/loser scenario working.

Creating a modal class to create a few modal elements actually seems pointless and a waste of time. I can just add the modals I need in the HTML and then use the hidden class via CSS to remove them when not in use.