4/4 Update
Got the functionality of the Ace working properly as far as I can tell. I've done a lot of testing and it seems to work as intended.

I made a "Play again" button work, even though it just re-instantiates the entire game... The way this is handled needs to be overhauled with an Event Loop.

Added the ability to store the Pot and Bankroll values next to the action buttons on the main play area.