4/5 Update
Bug fix:

When the original deck object was created there was no function to check to see if the deck had run out of cards. If the player had played long enough; the deck would run out and the game would crash due to there being nothing to deal to the players.

I created a function to correct this that assigns a fresh deck of cards to the old deck if it starts to run out (<10 cards in old deck)

Still needs fixing:

Due to testing, I haven't implemented a lose condition for the player. Currently they can go below 0 on their current bankroll. This is left in intentionally to test the game logic.